fixed freezing bug and changed price to long in backend

+++ MarketMakerAPI/Contracts/SignalR/NewOrderRequest.cs
@@ -2,7 +2,7 @@
 
 public record NewOrderRequest(
     string Market,
-    int Price,
+    long Price,
     int Quantity,
     string RequestReference 
 );
