+++ MarketMakerAPI/Contracts/SignalR/NewOrderResponse.cs
@@ -7,7 +7,7 @@ namespace MarketMaker.Contracts;
 public record NewOrderResponse(
     string? User,
     string Market,
-    int Price,
+    long Price,
     int Quantity,
     DateTime TimeStamp,
     Guid Id,
