+++ MarketMakerAPI/Models/Order.cs
@@ -3,7 +3,7 @@
 namespace MarketMaker.Models;
 
 [Serializable]
-public record Order(Guid Id, string User, int Price, string Market, int Quantity, DateTime TimeStamp)
+public record Order(Guid Id, string User, long Price, string Market, int Quantity, DateTime TimeStamp)
 {
     public int Quantity { get; set; } = Quantity;
 }
